Break out the polar terms to separate headers
Seems to help the compilation speed a little bit
